Refactor Command codes
