+++ src/main/java/seedu/quotesify/commands/AddCommand.java
@@ -6,6 +6,7 @@ import seedu.quotesify.book.BookList;
 import seedu.quotesify.category.Category;
 import seedu.quotesify.category.CategoryList;
 import seedu.quotesify.category.CategoryParser;
+import seedu.quotesify.commands.add.AddCategoryCommand;
 import seedu.quotesify.exception.QuotesifyException;
 import seedu.quotesify.lists.ListManager;
 import seedu.quotesify.quote.Quote;
@@ -28,10 +29,12 @@ import java.util.ArrayList;
 public class AddCommand extends Command {
     public static Logger addLogger = Logger.getLogger("QuotesifyLogger");
 
-    private String type;
-    private String information;
+    public String type;
+    public String information;
+    private String arguments;
 
     public AddCommand(String arguments) {
+        this.arguments = arguments;
         String[] details = arguments.split(" ", 2);
 
         // if user did not provide arguments, let details[1] be empty string
@@ -63,8 +66,7 @@ public class AddCommand extends Command {
             break;
         case TAG_CATEGORY:
             addLogger.log(Level.INFO, "going to add category to book/quote");
-            CategoryList categories = (CategoryList) ListManager.getList(ListManager.CATEGORY_LIST);
-            addCategoryToBookOrQuote(categories, ui);
+            new AddCategoryCommand(arguments).execute(ui, storage);
             break;
         case TAG_RATING:
             addLogger.log(Level.INFO, "going to add rating to book");
@@ -158,112 +160,6 @@ public class AddCommand extends Command {
             System.out.println(e.getMessage());
             addLogger.log(Level.INFO, "add quote to quote list failed");
             addLogger.log(Level.WARNING, e.getMessage());
-        }
-    }
-
-    private void addCategoryToBookOrQuote(CategoryList categories, TextUi ui) {
-        String[] tokens = information.split(" ");
-        String[] parameters = CategoryParser.getRequiredParameters(tokens);
-        int result = CategoryParser.validateParametersResult(parameters);
-        if (result == 1) {
-            executeParameters(categories, parameters, ui);
-        } else if (result == 0) {
-            ui.printErrorMessage(ERROR_MISSING_BOOK_OR_QUOTE);
-        } else {
-            ui.printErrorMessage(ERROR_MISSING_CATEGORY);
-        }
-    }
-
-    private void executeParameters(CategoryList categoryList, String[] parameters, TextUi ui) {
-        try {
-            String categoryNames = parameters[0];
-            assert !categoryNames.isEmpty() : "category name should not be empty";
-
-            List<String> categories = CategoryParser.parseCategoriesToList(categoryNames);
-            for (String categoryName : categories) {
-                addCategoryToList(categoryList, categoryName);
-                Category category = categoryList.getCategoryByName(categoryName);
-
-                String bookNum = parameters[1];
-                String quoteNum = parameters[2];
-
-                addCategoryToBook(category, bookNum, ui);
-                addCategoryToQuote(category, quoteNum, ui);
-                categoryList.updateListInCategory(category);
-            }
-        } catch (QuotesifyException e) {
-            addLogger.log(Level.WARNING, e.getMessage());
-            ui.printErrorMessage(e.getMessage());
-        }
-    }
-
-    private void addCategoryToList(CategoryList categories, String categoryName) {
-        if (!categories.isExistingCategory(categoryName)) {
-            categories.add(new Category(categoryName));
-        }
-    }
-
-    private void addCategoryToBook(Category category, String bookNum, TextUi ui) {
-        // ignore this action if user did not provide book title
-        if (bookNum.isEmpty()) {
-            return;
-        }
-
-        BookList bookList = (BookList) ListManager.getList(ListManager.BOOK_LIST);
-        try {
-            int bookIndex = Integer.parseInt(bookNum) - 1;
-            Book book = bookList.getBook(bookIndex);
-
-            if (book.getCategories().contains(category.getCategoryName())) {
-                String errorMessage = String.format(ERROR_CATEGORY_ALREADY_EXISTS,
-                        category.getCategoryName(), book.getTitle());
-                throw new QuotesifyException(errorMessage);
-            }
-
-            book.getCategories().add(category.getCategoryName());
-            ui.printAddCategoryToBook(book.getTitle(), category.getCategoryName());
-            addLogger.log(Level.INFO, "add category to book success");
-        } catch (IndexOutOfBoundsException e) {
-            addLogger.log(Level.WARNING, ERROR_NO_BOOK_FOUND);
-            ui.printErrorMessage(ERROR_NO_BOOK_FOUND);
-        } catch (NumberFormatException e) {
-            addLogger.log(Level.WARNING, ERROR_INVALID_BOOK_NUM);
-            ui.printErrorMessage(ERROR_INVALID_BOOK_NUM);
-        } catch (QuotesifyException e) {
-            addLogger.log(Level.WARNING, e.getMessage());
-            ui.printErrorMessage(e.getMessage());
-        }
-    }
-
-    private void addCategoryToQuote(Category category, String quoteNum, TextUi ui) {
-        // ignore this action if user did not provide quote number
-        if (quoteNum.isEmpty()) {
-            return;
-        }
-
-        QuoteList quoteList = (QuoteList) ListManager.getList(ListManager.QUOTE_LIST);
-        try {
-            int quoteIndex = Integer.parseInt(quoteNum) - 1;
-            Quote quote = quoteList.getList().get(quoteIndex);
-
-            if (quote.getCategories().contains(category.getCategoryName())) {
-                String errorMessage = String.format(ERROR_CATEGORY_ALREADY_EXISTS,
-                        category.getCategoryName(), quote.getQuote());
-                throw new QuotesifyException(errorMessage);
-            }
-
-            quote.getCategories().add(category.getCategoryName());
-            ui.printAddCategoryToQuote(quote.getQuote(), category.getCategoryName());
-            addLogger.log(Level.INFO, "add category to quote success");
-        } catch (IndexOutOfBoundsException e) {
-            addLogger.log(Level.WARNING, ERROR_NO_QUOTE_FOUND);
-            ui.printErrorMessage(ERROR_NO_QUOTE_FOUND);
-        } catch (NumberFormatException e) {
-            addLogger.log(Level.WARNING, ERROR_INVALID_QUOTE_NUM);
-            ui.printErrorMessage(ERROR_INVALID_QUOTE_NUM);
-        } catch (QuotesifyException e) {
-            addLogger.log(Level.WARNING, e.getMessage());
-            ui.printErrorMessage(e.getMessage());
         }
     }
